Separate pagination logic into its own module
